Week2/DailyProject/chatbot/requirements.txt
streamlit
groq
duckduckgo-search
ddgs
requests
beautifulsoup4
lxml
python-dotenv